Define function template in header file
